Praise be unto his glorious name, unknown
Oh, and also fixing all of the issues brought up in PR review.
