+++ Modules/galaxy/__init__.py
@@ -1,0 +1,15 @@
+"""
+__init__.py
+
+Copyright (c) 2018 The Fuel Rat Mischief,
+All rights reserved.
+
+Licensed under the BSD 3-Clause License.
+
+See LICENSE.md
+"""
+
+from .galaxy import Galaxy
+from .star_system import StarSystem
+
+__all__ = ["Galaxy", "StarSystem"]
